LayoutTests/platform/chromium/fast/forms/calendar-picker/date-picker-events.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../../../fast/js/resources/js-test-pre.js"></script>
<script src="../../../../../fast/forms/resources/picker-common.js"></script>
<script src="resources/calendar-picker-common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<input type="date" id="date1" value="2000-01-02">

<script>
description('Tests if value selection by calendar picker dispatches correct events.');

var eventsCounter = {};
function recordEvent(event) {
    if (eventsCounter[event.type] === undefined)
        eventsCounter[event.type] = 0;
    eventsCounter[event.type]++;
    debug('==> "' + event.type + '" event was dispatched.');
}

var date1 = document.getElementById('date1');
date1.addEventListener('input', recordEvent, false);
date1.addEventListener('change', recordEvent, false);

openPicker(date1, test1);

function test1() {
    eventSender.keyDown('rightArrow');
    debug('Choosing a new value from the calendar picker. "input" and "change" events should be dispatched in this order.');
    eventSender.keyDown('\n');
    shouldBeEqualToString('date1.value', '2000-01-03');

    shouldBe('eventsCounter.input', '1');
    shouldBe('eventsCounter.change', '1');

    eventsCounter = {};

    debug('Choosing the same value from the calendar picker. No events should be dispatched.');
    eventSender.keyDown('\n');
    shouldBeEqualToString('date1.value', '2000-01-03');
    shouldBeUndefined('eventsCounter.input');
    shouldBeUndefined('eventsCounter.change');

    finishJSTest();
}
</script>
<script src="../../../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
