LayoutTests/platform/chromium/fast/forms/calendar-picker/date-picker-events-expected.txt

Tests if value selection by calendar picker dispatches correct events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Choosing a new value from the calendar picker. "input" and "change" events should be dispatched in this order.
==> "input" event was dispatched.
==> "change" event was dispatched.
PASS date1.value is "2000-01-03"
PASS eventsCounter.input is 1
PASS eventsCounter.change is 1
Choosing the same value from the calendar picker. No events should be dispatched.
PASS date1.value is "2000-01-03"
PASS eventsCounter.input is undefined.
PASS eventsCounter.change is undefined.
PASS successfullyParsed is true

TEST COMPLETE
